+++ lib/src/main/java/com/smileidentity/util/Util.kt
@@ -24,9 +24,12 @@ import androidx.annotation.StringRes
 import androidx.camera.core.impl.utils.Exif
 import androidx.core.graphics.scale
 import com.google.mlkit.vision.common.InputImage
+import com.smileidentity.R
 import com.smileidentity.SmileID
 import com.smileidentity.SmileID.moshi
 import com.smileidentity.SmileIDCrashReporting
+import com.smileidentity.compose.consent.bvn.BvnOtpVerificationMode
+import com.smileidentity.models.BvnVerificationMode
 import com.smileidentity.models.SmileIDException
 import kotlinx.coroutines.CoroutineExceptionHandler
 import retrofit2.HttpException
@@ -34,6 +37,47 @@ import timber.log.Timber
 import java.io.File
 import java.io.Serializable
 import java.nio.ByteBuffer
+
+internal fun createBvnOtpVerificationModes(maps: List<BvnVerificationMode>):
+    List<BvnOtpVerificationMode> {
+    val verificationModes = mutableListOf<BvnOtpVerificationMode>()
+
+    for (map in maps) {
+        val modeEntry = map.entries.firstOrNull() // Assuming there's only one entry in each map
+        modeEntry?.let { entry ->
+            val mode = entry.key
+            val value = entry.value
+
+            val descriptionRes = when (mode) {
+                "sms" -> R.string.si_bvn_sms_verification
+                "email" -> R.string.si_bvn_email_verification
+                else -> R.string.si_bvn_sms_verification
+            }
+
+            val iconRes = when (mode) {
+                "sms" -> R.drawable.si_bvn_mode_sms
+                "email" -> R.drawable.si_bvn_mode_email
+                else -> R.drawable.si_bvn_mode_sms
+            }
+
+            val otpSentBy = when (mode) {
+                "sms" -> "sms"
+                "email" -> "email"
+                else -> ""
+            }
+
+            val verificationMode = BvnOtpVerificationMode(
+                mode = value,
+                otpSentBy = otpSentBy,
+                description = descriptionRes,
+                icon = iconRes,
+            )
+            verificationModes.add(verificationMode)
+        }
+    }
+
+    return verificationModes
+}
 
 internal fun Context.toast(@StringRes message: Int) {
     Toast.makeText(this, message, Toast.LENGTH_LONG).show()